third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/1.1/plugintypes-invalid-expected.txt

CONSOLE ERROR: 'plugin-types' Content Security Policy directive is empty; all plugins will be blocked. To disallow all plugins, the "object-src 'none'" directive should be used instead.

CONSOLE ERROR: line 16: Refused to load 'data:application/x-blink-test-plugin,' (MIME type 'application/x-blink-test-plugin') because it violates the following Content Security Policy Directive: 'plugin-types '.

CONSOLE ERROR: 'plugin-types' Content Security Policy directive is empty; all plugins will be blocked. To disallow all plugins, the "object-src 'none'" directive should be used instead.

CONSOLE ERROR: line 16: Refused to load 'data:application/x-blink-test-plugin,' (MIME type 'application/x-blink-test-plugin') because it violates the following Content Security Policy Directive: 'plugin-types '.

CONSOLE ERROR: Invalid plugin type in 'plugin-types' Content Security Policy directive: ''none''. Did you mean to set the object-src directive to 'none'?

CONSOLE ERROR: line 16: Refused to load 'data:application/x-blink-test-plugin,' (MIME type 'application/x-blink-test-plugin') because it violates the following Content Security Policy Directive: 'plugin-types 'none''.

CONSOLE ERROR: Invalid plugin type in 'plugin-types' Content Security Policy directive: 'text'.

CONSOLE ERROR: line 16: Refused to load 'data:application/x-blink-test-plugin,' (MIME type 'application/x-blink-test-plugin') because it violates the following Content Security Policy Directive: 'plugin-types text'.

CONSOLE ERROR: Invalid plugin type in 'plugin-types' Content Security Policy directive: 'text/'.

CONSOLE ERROR: line 16: Refused to load 'data:application/x-blink-test-plugin,' (MIME type 'application/x-blink-test-plugin') because it violates the following Content Security Policy Directive: 'plugin-types text/'.

CONSOLE ERROR: Invalid plugin type in 'plugin-types' Content Security Policy directive: '/text'.

CONSOLE ERROR: line 16: Refused to load 'data:application/x-blink-test-plugin,' (MIME type 'application/x-blink-test-plugin') because it violates the following Content Security Policy Directive: 'plugin-types /text'.

CONSOLE ERROR: Invalid plugin type in 'plugin-types' Content Security Policy directive: 'text//plain'.

CONSOLE ERROR: line 16: Refused to load 'data:application/x-blink-test-plugin,' (MIME type 'application/x-blink-test-plugin') because it violates the following Content Security Policy Directive: 'plugin-types text//plain'.

CONSOLE ERROR: Invalid plugin type in 'plugin-types' Content Security Policy directive: 'text/plainapplication/nospace'.

CONSOLE ERROR: line 16: Refused to load 'data:application/x-blink-test-plugin,' (MIME type 'application/x-blink-test-plugin') because it violates the following Content Security Policy Directive: 'plugin-types text/plainapplication/nospace'.

CONSOLE ERROR: Invalid plugin type in 'plugin-types' Content Security Policy directive: 'text'.

CONSOLE MESSAGE: Blink Test Plugin: initializing
This tests our handling of invalid `plugin-types` CSP directives. Consider this test passing if each of the following frames contains either "PASS" or no text at all.

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------


--------
Frame: '<!--framePath //<!--frame1-->-->'
--------


--------
Frame: '<!--framePath //<!--frame2-->-->'
--------


--------
Frame: '<!--framePath //<!--frame3-->-->'
--------


--------
Frame: '<!--framePath //<!--frame4-->-->'
--------


--------
Frame: '<!--framePath //<!--frame5-->-->'
--------


--------
Frame: '<!--framePath //<!--frame6-->-->'
--------


--------
Frame: '<!--framePath //<!--frame7-->-->'
--------


--------
Frame: '<!--framePath //<!--frame8-->-->'
--------
PASS.
